+++ packages/hapify-templates-rest-client/.hapifyrc.js
@@ -1,4 +1,0 @@
-module.exports = {
-  name: 'Hapify templates for Nestjs models',
-  description: 'Generate the models for a Nestjs application',
-};